CRUDs completos do usuario

+++ project/src/main/java/com/lumiere/project/Controllers/UsersControllers.java
@@ -1,4 +1,4 @@
-package com.lumiere.project.Controllers;
+package com.lumiere.project.controllers;
 
 import java.util.List;
 import java.util.Map;
@@ -26,6 +26,7 @@ import com.lumiere.project.dto.CadastrarDTO;
 import com.lumiere.project.dto.LoginDTO;
 import com.lumiere.project.dto.LoginResponseDTO;
 import com.lumiere.project.entities.UsersEntities;
+import com.lumiere.project.entities.VideosWorkshopEntities;
 import com.lumiere.project.infra.TokenService;
 import com.lumiere.project.repositories.UsersRepositories;
 
@@ -47,12 +48,9 @@ public class UsersControllers {
 	@CrossOrigin(origins = "http://localhost:4200")
 	@GetMapping("/buscarTodos")
 	public List<BuscarDadosUserDTO> listUsers() {
-		return repository.findAll().stream()
-				.map(u -> new BuscarDadosUserDTO(u.getIdUser().toString(), u.getNome(), u.getEmail(), u.getTelefone(), u.getCpf()))
-				.toList();
+		return repository.findAll().stream().map(u -> new BuscarDadosUserDTO(u.getIdUser().toString(), u.getNome(),
+				u.getEmail(), u.getTelefone(), u.getCpf())).toList();
 	}
-	
-
 
 	// cadastrar usuario padrão
 	@CrossOrigin(origins = "http://localhost:4200")
@@ -76,13 +74,15 @@ public class UsersControllers {
 			var auth = this.authenticationManager.authenticate(usernamePassword);
 			// geramos o token JWT para o usuário autenticado.
 			var token = tokenService.generateToken((UsersEntities) auth.getPrincipal());
-			
-			var usuarioAutenticado = (UsersEntities) auth.getPrincipal();
-			
-			//UsersEntities u = null;
-			BuscarDadosUserDTO dados = new BuscarDadosUserDTO(usuarioAutenticado.getIdUser().toString(), usuarioAutenticado.getNome(), usuarioAutenticado.getEmail(), usuarioAutenticado.getTelefone(), usuarioAutenticado.getCpf());
 
-			 LoginResponseDTO response = new LoginResponseDTO(token, dados);
+			var usuarioAutenticado = (UsersEntities) auth.getPrincipal();
+
+			// UsersEntities u = null;
+			BuscarDadosUserDTO dados = new BuscarDadosUserDTO(usuarioAutenticado.getIdUser().toString(),
+					usuarioAutenticado.getNome(), usuarioAutenticado.getEmail(), usuarioAutenticado.getTelefone(),
+					usuarioAutenticado.getCpf());
+
+			LoginResponseDTO response = new LoginResponseDTO(token, dados);
 			// 4. Retornamos o status 200 OK com o token no corpo da resposta.
 			return ResponseEntity.ok().body(response);
 
@@ -94,28 +94,51 @@ public class UsersControllers {
 
 	@CrossOrigin(origins = "http://localhost:4200")
 	@DeleteMapping("/{id}")
-	public String deletarUsuario(@PathVariable String id) {
+	public ResponseEntity<?> deletarUsuario(@PathVariable String id) {
 		try {
+			Optional<UsersEntities> userId = repository.findById(id);
+			if (userId.isEmpty()) {
+				return ResponseEntity.badRequest().body(Map.of("error", "usuario não encontrado no banco de dados"));
+			}
 			repository.deleteById(id);
-			return "Usuario deletado com sucesso";
+			return ResponseEntity.ok().body(Map.of("sucess", "usuario deletado"));
 		} catch (Exception e) {
-			return "não foi possível deletar o usuario " + e.getMessage();
+			return ResponseEntity.badRequest().body(Map.of("error", "Erro inesperado"));
 		}
 	}
-	
-	@CrossOrigin(origins = "http://localhost:4200")
-	@PutMapping("/atualizarEmail/{idUser}")
-	public ResponseEntity atualizarEmail(@PathVariable String idUser, @RequestBody CadastrarDTO curso) {
-		Optional<UsersEntities> optionalUser = repository.findById(idUser); // o optional sempre verifica se existe
-																				// ou n
-		if (optionalUser.isEmpty()) {
-			return ResponseEntity.badRequest().build();
-		}
-		UsersEntities userEmail = optionalUser.get();
-		userEmail.setEmail(curso.email());
-		repository.save(userEmail);
 
-		return ResponseEntity.ok(userEmail);
+	@CrossOrigin(origins = "http://localhost:4200")
+	@PutMapping("/atualizar/{idUser}")
+	public ResponseEntity atualizarEmail(@PathVariable String idUser, @RequestBody CadastrarDTO usuario) {
+		try {
+			
+		Optional<UsersEntities> optionalUser = repository.findById(idUser); // o optional sempre
+																								// verifica se existe
+		// ou n
+		if (optionalUser.isEmpty()) {
+			return ResponseEntity.badRequest().body(Map.of("error", "não foi possívela achar o aluno no banco de dados"));
+		}
+
+		UsersEntities user = optionalUser.get();
+
+		if (usuario.email() != null && !usuario.email().isBlank()) {
+			user.setEmail(usuario.email());
+		}
+
+		if (usuario.telefone() != null && !usuario.telefone().isBlank()) {
+			user.setTelefone(usuario.telefone());
+		}
+
+		if (usuario.nome() != null && !usuario.nome().isBlank()) {
+			user.setNome(usuario.nome());
+		}
+
+		repository.save(user);
+
+		return ResponseEntity.ok().body(Map.of("sucesso", "Aluno seus dados foram atualizados"));
+		} catch (Exception e) {
+			return ResponseEntity.badRequest().body(Map.of("error", "Erro inesperado"));
+		}
 	}
 
 }
